Discord Bot 1.04

+++ index.js
@@ -1,6 +1,6 @@
 const Discord = require('discord.js');
 const bot = new Discord.Client();
-const token = '';
+const token = 'NTczNjAxNzU0MjE4ODg5MjM2.XMxhug.W4JbRFfXnEasCuYuVeUXb9Jr8y0';
 const PREFIX = '~';
 
 bot.on('ready', () => {
@@ -11,16 +11,21 @@ bot.on('message', message => {
     let args = message.content.substring(PREFIX.length).split(" ");
 
     switch(args[0]){
+
+        case 'help':
+            if(args[0] && !args[1]) message.channel.send(" ``` The Prefix for Commands is ! \n !Ping - Pong \n !delete [arg] - Deletes messages based on given integer arguement ```");
+            if(args[1] == 1) message.channel.send(" ``` !Reminder [arg] [arg] [arg] [arg] - Follows Day Month Year Time format. Time is in military. ```");
+            if(args[1] == 2) message.channel.send(" ``` !Math [arg] [arg] [arg] - First argument is operation. 1 for addition. 2 for multiplication. 3 for division. 4 for subraction. Args 2 and 3 are varaibles A and B ```");
+            break;
+
+            /**
+            * Ping
+            * Pong
+            */
+
         case 'ping':
-            message.channel.sendMessage("pong!");
+            message.channel.send("pong!");
             break;
-        case 'Eddie':
-            message.channel.sendMessage("I don't really care");
-            break;
-        case 'Randy':
-            message.channel.sendMessage("Your mother!");
-            break;
-        case 'purge':
 
             /**
              * Delete
@@ -29,6 +34,7 @@ bot.on('message', message => {
             
         case 'delete':
             if(!args[1]) return message.reply('Define total number of lines');
+            if(args[1] <= 0) return message.reply('Please insert a value greater than zero');
             message.channel.bulkDelete(args[1]);
             break;
 
@@ -54,14 +60,14 @@ bot.on('message', message => {
              * Takes in three arguements -> (Operation, Value A, Value B)
              */
 
-            case 'math':
+        case 'math':
             if(!args[3]) return message.replay("Requires three arguements, 1 - Addition, 2 - Multiplication, 3 - Division, 4 - Subtraction. Then the two values, A and B");
             var operation = args[1];
-            var a = arg[2];
-            var b = arg[3];
+            var a = args[2];
+            var b = args[3];
 
             if(args[1] == 1){
-                return message.channel.send(a+b);
+               return message.channel.send(a+b);
             }
             if(args[1] == 2){
                 return message.channel.send(a*b);
